requirements.txt
requests
psycopg2-binary
redis
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
